+++ lib/src/affix.dart
@@ -14,8 +14,8 @@ class Affix extends Base {
    * 
    */
   Affix(Element element, {int offsetTop(), int offsetBottom()}) : 
-  this.offsetTop = p.fallback(offsetTop, () => () => _DEFAULT_OFFSET),
-  this.offsetBottom = p.fallback(offsetBottom, () => () => _DEFAULT_OFFSET),
+  this.offsetTop = offsetTop ?? (() => _DEFAULT_OFFSET),
+  this.offsetBottom = offsetBottom ?? (() => _DEFAULT_OFFSET),
   super(element, _NAME) {
     $(window)
     ..on('scroll.affix.data-api', (QueryEvent e) => checkPosition())
@@ -29,7 +29,7 @@ class Affix extends Base {
    * 
    */
   static Affix wire(Element element, [Affix create()]) =>
-      p.wire(element, _NAME, p.fallback(create, () => () => new Affix(element)));
+      p.wire(element, _NAME, create ?? (() => new Affix(element)));
   
   /**
    * 
@@ -38,8 +38,8 @@ class Affix extends Base {
     if (p.isHidden(element))
       return;
     
-    final int offsetTop = p.fallback(this.offsetTop(), () => _DEFAULT_OFFSET);
-    final int offsetBottom = p.fallback(this.offsetBottom(), () => _DEFAULT_OFFSET);
+    final int offsetTop = this.offsetTop() ?? _DEFAULT_OFFSET;
+    final int offsetBottom = this.offsetBottom() ?? _DEFAULT_OFFSET;
     
     final int scrollHeight = $document().height;
     final int scrollTop = window.pageYOffset;